move TimeOutChecker into its own file
