+++ pkg/subscription/handler.go
@@ -198,29 +198,3 @@ func (u *UniversalProtocolHandler) Handle(ctx context.Context) {
 		}
 	}
 }
-
-type TimeOutParams struct {
-	Name            string
-	Logger          abstractlogger.Logger
-	TimeOutContext  context.Context
-	TimeOutAction   func()
-	TimeOutDuration time.Duration
-}
-
-func TimeOutChecker(params TimeOutParams) {
-	timer := time.NewTimer(params.TimeOutDuration)
-	defer timer.Stop()
-
-	for {
-		select {
-		case <-params.TimeOutContext.Done():
-			return
-		case <-timer.C:
-			params.Logger.Error("time out happened",
-				abstractlogger.String("name", params.Name),
-			)
-			params.TimeOutAction()
-			return
-		}
-	}
-}
